requirements.txt
requests==2.32.3
streamlit==1.37.1
openai==1.42.0
pandas==2.2.2
matplotlib==3.9.2
PyPDF2==3.0.1
pillow==10.4.0